% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_annotation.R
\name{Annotator.annotate}
\alias{Annotator.annotate}
\title{Annotate a data table/frame with additional fields.}
\usage{
Annotator.annotate(
  records,
  fields,
  include_errors = FALSE,
  raw = FALSE,
  env = get_connection()
)
}
\arguments{
\item{records}{The data to annotate as a data frame.}

\item{fields}{The fields to add.}

\item{include_errors}{whether to include errors in the output.}

\item{raw}{whether to return the raw response.}

\item{env}{Custom client environment.}
}
\description{
Annotate a data table/frame with additional fields.
}
