% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_application.R
\name{Application.create}
\alias{Application.create}
\title{Application.create}
\usage{
Application.create(name, redirect_uris, env = get_connection(), ...)
}
\arguments{
\item{name}{The name of the application.}

\item{redirect_uris}{A list of space-separated OAuth2 redirect URIs.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional application attributes.}
}
\description{
Create a new QuartzBio EDP application.
}
\examples{
\dontrun{
Application.create(
    name = "My new application",
    redirect_uris = "http://localhost:3838/"
)
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{solvebio_api}
