% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{autoconnect}
\alias{autoconnect}
\title{tries to connect, using environment variables or the default profile}
\usage{
autoconnect(check = FALSE)
}
\arguments{
\item{check}{whether to check the connection, mostly for debugging purposes}
}
\value{
the connection
}
\description{
tries to connect, using environment variables or the default profile
}
\seealso{
connect

read_connection_profile
}
