% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_beacon.R
\name{Beacon.all}
\alias{Beacon.all}
\title{Beacon.all}
\usage{
Beacon.all(env = get_connection(), ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters.}
}
\description{
Retrieves the metadata about all beacons on QuartzBio EDP accessible to the current user.
}
\examples{
\dontrun{
Beacon.all()
}

}
\concept{quartzbio_api}
