% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_beacon_set.R
\name{BeaconSet.delete}
\alias{BeaconSet.delete}
\title{BeaconSet.delete}
\usage{
BeaconSet.delete(id, env = get_connection())
}
\arguments{
\item{id}{The ID of the beacon set.}

\item{env}{(optional) Custom client environment.}
}
\description{
Delete a specific beacon set (including all its beacons) from QuartzBio EDP.
}
\examples{
\dontrun{
BeaconSet.delete("1234")
}

}
\concept{quartzbio_api}
