% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_beacon_set.R
\name{BeaconSet.retrieve}
\alias{BeaconSet.retrieve}
\title{BeaconSet.retrieve}
\usage{
BeaconSet.retrieve(id, env = get_connection())
}
\arguments{
\item{id}{The ID of the beacon set.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific beacon set on QuartzBio EDP.
}
\examples{
\dontrun{
BeaconSet.retrieve("1234")
}

}
\references{
\url{https://docs.solvebio.com/}
}
