% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_setup.R
\name{configure_logger}
\alias{configure_logger}
\title{Configure log file and set the log-level}
\usage{
configure_logger(log_file, log_level = "INFO")
}
\arguments{
\item{log_file}{Path to the file to log records}

\item{log_level}{Set log level. Default is set to INFO. Available levels: FATAL, ERROR, WARN, INFO, DEBUG}
}
\value{
The log file with set log level threshold id logging file is configured successfully.
\code{"console"} if logging defaults to console or \code{NULL}if logging is disabled. (due to invalid path or missing logger package)
}
\description{
Setup the  log record appender function for the file provided
}
\concept{quartzbio_api}
