% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{connect to the QuartzBio EDP API and return the connection}
\usage{
connect(
  secret = get_env("EDP_API_SECRET", get_env("SOLVEBIO_ACCESS_TOKEN",
    get_env("SOLVEBIO_API_KEY"))),
  host = get_env("EDP_API_HOST", get_env("SOLVEBIO_API_HOST", EDP_DEFAULT_API_HOST)),
  check = TRUE
)
}
\arguments{
\item{secret}{a QuartzBio EDP \strong{API key}  or \strong{token} as a string.
Defaults to the \code{EDP_API_SECRET} environment variable if set,
otherwise to the legacy \code{SOLVEBIO_ACCESS_TOKEN} var, then to
to the \code{SOLVEBIO_API_KEY} var.}

\item{host}{the QuartzBio EDP \strong{API host} as a string.
Defaults to the \code{EDP_API_HOST} environment variable if set, otherwise to the
legacy \code{SOLVEBIO_API_HOST} var.}

\item{check}{whether to check the connection, mostly for debugging purposes}
}
\value{
a connection object
}
\description{
connect to the QuartzBio EDP API and return the connection
}
\examples{
\dontrun{
#  using API key
conn <- connect("MYKEY")
# using env vars
conn <- connect()
# using token and explicit host
conn <- connect("MYTOKEN", "https://xxxx.yy.com")
}

}
