% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_with_profile}
\alias{connect_with_profile}
\title{connect to the QuartzBio EDP API using a saved profile}
\usage{
connect_with_profile(..., check = TRUE)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=read_connection_profile]{read_connection_profile}}
  \describe{
    \item{\code{profile}}{the name of the profile, as a string}
    \item{\code{path}}{the path to the connection profiles file, as a string}
  }}

\item{check}{whether to check the connection, mostly for debugging purposes}
}
\description{
connect to the QuartzBio EDP API using a saved profile
}
