% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{create_model_df_from_fields}
\alias{create_model_df_from_fields}
\title{create a model data frame from a list of column names and some meta data fields.}
\usage{
create_model_df_from_fields(cols, fields, titles = TRUE, ordering = TRUE)
}
\arguments{
\item{cols}{list of column names}

\item{fields}{meta data fields from EDP dataset}

\item{titles}{Set the dataframe column names as per titles. Default is TRUE}

\item{ordering}{Set the order of columns based on ordering. Default is TRUE}
}
\value{
A model empty R dataframe
}
\description{
Creates a model empty data frame from a list of column names and meta data fields
of the EDP dataset.
}
