% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{createEnv}
\alias{createEnv}
\title{createEnv}
\usage{
createEnv(token, token_type = "Token", host = .config$host)
}
\arguments{
\item{token}{A QuartzBio EDP API key or OAuth2 token}

\item{token_type}{QuartzBio EDP token type (default: Token)}

\item{host}{(optional) The QuartzBio EDP API host (default: https://api.solvebio.com)}
}
\description{
Create a new QuartzBio EDP environment.
}
\examples{
\dontrun{
env <- createEnv("MyAPIkey")
User.retrieve(env = myEnv)
}

}
\references{
\url{https://docs.solvebio.com/}
}
