% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset.R
\name{Dataset.all}
\alias{Dataset.all}
\title{Retrieves the metadata about datasets on QuartzBio EDP.}
\usage{
Dataset.all(env = get_connection(), ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. page).}
}
\description{
Retrieves the metadata about datasets on QuartzBio EDP.
}
\examples{
\dontrun{
Dataset.all()
}

}
\concept{quartzbio_api}
