% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset_create}
\alias{Dataset_create}
\title{creates a new Dataset.}
\usage{
Dataset_create(
  vault_id,
  vault_path,
  description = NULL,
  metadata = NULL,
  tags = NULL,
  storage_class = NULL,
  capacity = NULL,
  conn = get_connection()
)
}
\arguments{
\item{vault_id}{a Vault ID as a string (e.g. "19").}

\item{vault_path}{a Vault path, as a string (e.g. "/d1/d2/foo.csv")}

\item{description}{the description as a string.}

\item{metadata}{metadata as a named list.}

\item{tags}{a list of tags as a character vector.}

\item{storage_class}{The Storage class of the vault \verb{('Standard', 'Standard-IA', 'Essential', 'Temporary', 'Performance', 'Archive')} as a string.}

\item{capacity}{The dataset capacity level (small, medium, or large).}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\description{
creates a new Dataset.
}
