% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_export_to_parquet}
\alias{dataset_export_to_parquet}
\title{Initiate a Dataset export to parquet}
\usage{
dataset_export_to_parquet(id = NULL, full_path = NULL, conn = get_connection())
}
\arguments{
\item{id}{(character) The ID of a QuartzBio EDP dataset, or a Dataset object.}

\item{full_path}{(character) a valid dataset full path, including the account, vault and path to EDP Dataset.}

\item{conn}{(optional) Custom client environment.}
}
\value{
URL to the parquet file for the given EDP dataset. If there is an
failure, an error will be raised.
}
\description{
Create a dataset export to parquet file and return the URL to export file
}
\concept{solvebio_api}
