% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset_import}
\alias{Dataset_import}
\title{imports data into an existing dataset}
\usage{
Dataset_import(
  dataset_id,
  commit_mode = NULL,
  records = NULL,
  df = NULL,
  target_fields = infer_fields_from_df(df),
  sync = FALSE,
  conn = get_connection(),
  ...
)
}
\arguments{
\item{dataset_id}{a Dataset ID as a string}

\item{commit_mode}{There are four commit modes that can be selected depending on the scenario:
append (default), overwrite, upsert, and delete.}

\item{records}{The data to annotate as a data frame.}

\item{df}{The data to import as a data.frame}

\item{target_fields}{A list of valid dataset fields to create or override in the import,
as a character vector.}

\item{sync}{whether to proceed in synchronous mode, i.e to wait for all
sub tasks to finish before returning.}

\item{conn}{a EDP connection object (as a named list or environment)}

\item{...}{passed to \code{\link[=Task_wait_for_completion]{Task_wait_for_completion()}}}
}
\description{
imports data into an existing dataset
}
