% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset_load}
\alias{Dataset_load}
\title{Dataset_load}
\usage{
Dataset_load(id = NULL, full_path = NULL, select_fields = NULL, ...)
}
\arguments{
\item{id}{The ID of a QuartzBio EDP dataset.}

\item{full_path}{a valid dataset full path, including the account, vault and path to EDP Dataset.}

\item{select_fields}{A character vector of field names to select in the results.
Tidy specification of fields can also be used. Check \code{\link[arrow:read_parquet]{arrow::read_parquet()}}}

\item{...}{Additional arguments passed to \code{ParquetFileReader$create()}}
}
\value{
A \code{tibble} which is the default, or an Arrow Table otherwise.
}
\description{
Loads large Quartzbio EDP dataset and returns an R data frame containing all records.
}
\concept{solvebio_api}
