% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset_query}
\alias{Dataset_query}
\title{queries data into a dataset.}
\usage{
Dataset_query(
  dataset_id,
  filters = NULL,
  facets = NULL,
  fields = NULL,
  exclude_fields = c("_id", "_commit"),
  ordering = NULL,
  query = NULL,
  limit = 10000,
  offset = NULL,
  all = FALSE,
  meta = TRUE,
  parallel = FALSE,
  workers = 4,
  conn = get_connection(),
  ...
)
}
\arguments{
\item{dataset_id}{a Dataset ID as a string}

\item{filters}{a filter expression as a JSON string.}

\item{facets}{A valid facets objects.}

\item{fields}{The fields to add.}

\item{exclude_fields}{A list of fields to exclude in the results, as a character vector.}

\item{ordering}{A list of fields to order/sort the results with, as a character vector.}

\item{query}{a string that matches any objects whose path contains that string.}

\item{limit}{The maximum number of elements to fetch, as an integer.
See also \code{page}.}

\item{offset}{the file offset (starts from 0).}

\item{all}{whether to fetch all data, by iterating if needed.}

\item{meta}{whether to retrieve fields meta data information to properly format, reorder
and rename the data frame}

\item{parallel}{whether to parallelize the API calls.}

\item{workers}{in parallel mode, the number of concurrent requests to make}

\item{conn}{a EDP connection object (as a named list or environment)}

\item{...}{Additional query parameters, passed to .request().}
}
\description{
queries data into a dataset.
}
