% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset_commit.R
\name{DatasetCommit.delete}
\alias{DatasetCommit.delete}
\title{DatasetCommit.delete}
\usage{
DatasetCommit.delete(id, env = get_connection())
}
\arguments{
\item{id}{String The ID or full name of a QuartzBio EDP dataset commit.}

\item{env}{(optional) Custom client environment.}
}
\description{
Deletes a specific dataset commit on QuartzBio EDP.
}
\examples{
\dontrun{
DatasetCommit.delete(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{solvebio_api}
