% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset_commit.R
\name{DatasetCommit.retrieve}
\alias{DatasetCommit.retrieve}
\title{DatasetCommit.retrieve}
\usage{
DatasetCommit.retrieve(id, env = get_connection())
}
\arguments{
\item{id}{String The ID of a QuartzBio EDP dataset commit.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific dataset commit on QuartzBio EDP.
}
\examples{
\dontrun{
DatasetCommit.retrieve(12345)
}

}
\concept{quartzbio_api}
