% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset_export.R
\name{DatasetExport.delete}
\alias{DatasetExport.delete}
\title{DatasetExport.delete}
\usage{
DatasetExport.delete(id, env = get_connection())
}
\arguments{
\item{id}{String The ID of a QuartzBio EDP dataset export.}

\item{env}{(optional) Custom client environment.}
}
\description{
Deletes a specific dataset export on QuartzBio EDP.
}
\examples{
\dontrun{
DatasetExport.delete(12345)
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{quartzbio_api}
