% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{DatasetField}
\alias{DatasetField}
\title{fetches a field metadata of a dataset by ID or (datasetid, field_name)}
\usage{
DatasetField(
  field_id = NULL,
  dataset_id = NULL,
  name = NULL,
  conn = get_connection()
)
}
\arguments{
\item{field_id}{a Field object ID.}

\item{dataset_id}{a Dataset ID as a string}

\item{name}{The name of the field}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
a DatasetField object
}
\description{
fetches a field metadata of a dataset by ID or (datasetid, field_name)
}
