% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{DatasetField_update}
\alias{DatasetField_update}
\title{updates an existing Dataset Field.}
\usage{
DatasetField_update(
  field_id,
  data_type = NULL,
  title = NULL,
  description = NULL,
  ordering = NULL,
  entity_type = NULL,
  expression = NULL,
  is_hidden = NULL,
  is_list = NULL,
  url_template = NULL,
  conn = get_connection()
)
}
\arguments{
\item{field_id}{a Field object ID.}

\item{data_type}{the data type. one of:
\itemize{
\item auto (the default)
\item boolean - Either True, False, or null
\item date - A string in ISO 8601 format, for example: \verb{2017-03-29T14:52:01}
\item double - A double-precision 64-bit IEEE 754 floating point.
\item float - single-precision 32-bit IEEE 754 floating point.
\item integer     A signed 32-bit integer with a minimum value of -231 and a maximum value of 231-1.
\item long     A signed 64-bit integer with a minimum value of -263 and a maximum value of 263-1.
\item object     A key/value, JSON-like object, similar to a Python dictionary.
\item string     A valid UTF-8 string up to 32,766 characters in length.
\item text     A valid UTF-8 string of any length, indexed for full-text search.
\item blob     A valid UTF-8 string of any length, not indexed for search.
}}

\item{title}{The field's display name, shown in the UI and in CSV/Excel exports, as a string}

\item{description}{the description as a string.}

\item{ordering}{A list of fields to order/sort the results with, as a character vector.}

\item{entity_type}{A valid entity type:
\itemize{
\item dataset - a Dataset ID (510110013133189334)
\item gene - A gene (EGFR)
\item genomic_region - A genomic region (GRCH38-7-55019017-55211628)
\item literature - A PubMed ID (19915526)
\item sample - A sample identifier (TCGA-02-0001)
\item variant - A genomic variant (GRCH38-7-55181378-55181378-T)
}}

\item{expression}{EDP xpressions are Python-like formulas that can be used to pull
data from datasets, calculate statistics, or run advanced algorithms.}

\item{is_hidden}{whether the field should be hidden from the UI.}

\item{is_list}{whether the result is expected to be a list.}

\item{url_template}{A URL template with one or more "value" sections that will be
interpolated with the field value and displayed as a link
in the dataset table.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
a DatasetField object
}
\description{
updates an existing Dataset Field.
}
