% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{DatasetFields}
\alias{DatasetFields}
\title{fetches the fields of a dataset.}
\usage{
DatasetFields(
  dataset_id,
  limit = NULL,
  page = NULL,
  all = FALSE,
  conn = get_connection(),
  ...
)
}
\arguments{
\item{dataset_id}{a Dataset ID as a string}

\item{limit}{The maximum number of elements to fetch, as an integer.
See also \code{page}.}

\item{page}{The number of the page to fetch, as an integer. starts from 1.
See also \code{limit}.}

\item{all}{whether to fetch all data, by iterating if needed.}

\item{conn}{a EDP connection object (as a named list or environment)}

\item{...}{Additional query parameters, passed to .request().}
}
\value{
a DatasetFieldList object
}
\description{
fetches the fields of a dataset.
}
\references{
\url{https://docs.solvebio.com/}
}
