% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset_import.R
\name{DatasetImport.delete}
\alias{DatasetImport.delete}
\title{DatasetImport.delete}
\usage{
DatasetImport.delete(id, env = get_connection())
}
\arguments{
\item{id}{String The ID of a QuartzBio EDP dataset import.}

\item{env}{(optional) Custom client environment.}
}
\description{
Deletes a specific dataset import on QuartzBio EDP.
}
\examples{
\dontrun{
DatasetImport.delete(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{solvebio_api}
