% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset_migration.R
\name{DatasetMigration.delete}
\alias{DatasetMigration.delete}
\title{DatasetMigration.delete}
\usage{
DatasetMigration.delete(id, env = get_connection())
}
\arguments{
\item{id}{String The ID of a QuartzBio EDP dataset migration.}

\item{env}{(optional) Custom client environment.}
}
\description{
Deletes specific dataset migration on QuartzBio EDP.
}
\examples{
\dontrun{
DatasetMigration.delete(12345)
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{quartzbio_api}
