% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset_template.R
\name{DatasetTemplate.create}
\alias{DatasetTemplate.create}
\title{DatasetTemplate.create}
\usage{
DatasetTemplate.create(env = get_connection(), ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Dataset template attributes.}
}
\description{
Create a QuartzBio EDP dataset template.
}
\examples{
\dontrun{
DatasetTemplate.create(name = "My Dataset Template")
}

}
\concept{quartzbio_api}
