% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset_template.R
\name{DatasetTemplate.delete}
\alias{DatasetTemplate.delete}
\title{DatasetTemplate.delete}
\usage{
DatasetTemplate.delete(id, env = get_connection())
}
\arguments{
\item{id}{String The ID or full name of a QuartzBio EDP dataset template.}

\item{env}{(optional) Custom client environment.}
}
\description{
Deletes a specific dataset template on QuartzBio EDP.
}
\examples{
\dontrun{
DatasetTemplate.delete(12345)
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{quartzbio_api}
