% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset_template.R
\name{DatasetTemplate.retrieve}
\alias{DatasetTemplate.retrieve}
\title{DatasetTemplate.retrieve}
\usage{
DatasetTemplate.retrieve(id, env = get_connection())
}
\arguments{
\item{id}{String The ID of a QuartzBio EDP dataset template.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific dataset template on QuartzBio EDP.
}
\examples{
\dontrun{
DatasetTemplate.retrieve(12345)
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{solvebio_api}
