% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_setup.R
\name{edp_health_check}
\alias{edp_health_check}
\title{edp_health_check}
\usage{
edp_health_check(get_vault_list = FALSE)
}
\arguments{
\item{get_vault_list}{Get the vault list created by the user}
}
\description{
Provides quick health check to test your EDP connection and retrieves user details and vaults
}
