% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{Expression.evaluate}
\alias{Expression.evaluate}
\title{Evaluate a QuartzBio EDP expression.}
\usage{
Expression.evaluate(
  expression,
  data_type = "string",
  is_list = FALSE,
  data = NULL,
  raw = FALSE,
  env = quartzbio.edp:::.config
)
}
\arguments{
\item{expression}{The EDP expression string.}

\item{data_type}{The data type to cast the expression result.}

\item{is_list}{whether the result is expected to be a list.}

\item{raw}{whether to return the raw response.}

\item{env}{Custom client environment.}
}
\description{
Evaluate a QuartzBio EDP expression.
}
\examples{
\dontrun{
Expression.evaluate("1 + 1", data_type = "integer", is_list = FALSE)
}
}
