% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pager.R
\name{fetch_all}
\alias{fetch_all}
\title{fetch all the pages for a possibly incomplete paginated API result}
\usage{
fetch_all(x, ..., parallel = FALSE, workers = 4)
}
\arguments{
\item{x}{an API result}

\item{...}{passed to future.apply::future_lapply()}
}
\value{
the object resulting in combining the current object/page and all subsequent pages
}
\description{
fetch all the pages for a possibly incomplete paginated API result
}
