% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pager.R
\name{fetch_next}
\alias{fetch_next}
\title{fetch the next page of data if any}
\usage{
fetch_next(x)
}
\arguments{
\item{x}{an API result}
}
\value{
the next page of data, or NULL if none
}
\description{
fetch the next page of data if any
}
