% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pager.R
\name{fetch_prev}
\alias{fetch_prev}
\title{fetch the previous page of data if any}
\usage{
fetch_prev(x)
}
\arguments{
\item{x}{an API result}
}
\value{
the previous page of data, or NULL if none
}
\description{
fetch the previous page of data if any
}
