% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{File_download}
\alias{File_download}
\title{utility function that downloads an EDP File into a local file}
\usage{
File_download(file_id, local_path, conn = get_connection(), ...)
}
\arguments{
\item{file_id}{a file Object ID.}

\item{local_path}{the path of a local file.}

\item{conn}{a EDP connection object (as a named list or environment)}

\item{...}{passed to File_download_content()}
}
\value{
the response
}
\description{
utility function that downloads an EDP File into a local file
}
