% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{File_get_download_url}
\alias{File_get_download_url}
\title{fetches the download URL of a file.}
\usage{
File_get_download_url(file_id, conn = get_connection())
}
\arguments{
\item{file_id}{a file Object ID.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
the download URL as a string
}
\description{
This URL can then be used to download the file using any HTTP client, such as
utils::download.file()
}
