% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{File_query}
\alias{File_query}
\title{queries the content of a file.}
\usage{
File_query(
  id,
  filters = NULL,
  limit = 10000,
  offset = NULL,
  all = FALSE,
  conn = get_connection()
)
}
\arguments{
\item{id}{a File ID}

\item{filters}{a filter expression as a JSON string.}

\item{limit}{The maximum number of elements to fetch, as an integer.
See also \code{page}.}

\item{offset}{the file offset (starts from 0).}

\item{all}{whether to fetch all data, by iterating if needed.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\description{
The file has to be parsable by EDP. Otherwise you can use \code{\link[=File_download]{File_download()}}
}
