% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{File_read}
\alias{File_read}
\title{convenience function to download a file into memory, just a wrapper over File_download()}
\usage{
File_read(file_id, local_path = tempfile(), conn = get_connection(), ...)
}
\arguments{
\item{file_id}{a file Object ID.}

\item{local_path}{where to download the file. Used for testing purposes.}

\item{conn}{a EDP connection object (as a named list or environment)}

\item{...}{passed to File_download()}
}
\value{
the file content
}
\description{
convenience function to download a file into memory, just a wrapper over File_download()
}
