% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{File_upload}
\alias{File_upload}
\title{uploads a file}
\usage{
File_upload(
  vault_id,
  local_path,
  vault_path,
  mimetype = mime::guess_type(local_path),
  conn = get_connection()
)
}
\arguments{
\item{vault_id}{a Vault ID as a string (e.g. "19").}

\item{local_path}{the path of a local file.}

\item{vault_path}{a Vault path, as a string (e.g. "/d1/d2/foo.csv")}

\item{mimetype}{the MIME type of the Object.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\description{
uploads a file
}
