% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\title{parses the maths-like syntax of filters.}
\usage{
filters(x)
}
\arguments{
\item{x}{the filters string to parse}
}
\value{
the download URL as a string
}
\description{
parses the maths-like syntax of filters.
}
