% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{Folder}
\alias{Folder}
\title{fetches a folder by id, full_path or (vault_id, path)}
\usage{
Folder(
  id = NULL,
  full_path = NULL,
  path = NULL,
  vault_id = NULL,
  conn = get_connection()
)
}
\arguments{
\item{id}{the folder ID.}

\item{full_path}{an Object full path, including the account, vault and path.}

\item{path}{the path of an object, with the folders (e.g. "/d1/d2/foo.csv").}

\item{vault_id}{a Vault ID as a string (e.g. "19").}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\description{
fetches a folder by id, full_path or (vault_id, path)
}
