% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{Folder_create}
\alias{Folder_create}
\title{creates  a folder.}
\usage{
Folder_create(
  vault_id,
  path,
  recursive = TRUE,
  parent_folder_id = NULL,
  conn = get_connection()
)
}
\arguments{
\item{vault_id}{a Vault ID as a string (e.g. "19").}

\item{path}{the folder path to create}

\item{recursive}{whether to recursively create the parent folders if they do not exist.}

\item{parent_folder_id}{the ID of the parent folder.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
the folder as an Object
}
\description{
creates  a folder.
}
