% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{format_df_like_model}
\alias{format_df_like_model}
\title{Format a data frame like the model data frame}
\usage{
format_df_like_model(df, model)
}
\arguments{
\item{df}{dataframe to be formatted}

\item{model}{model dataframe}
}
\value{
formatted dataframe
}
\description{
format the dataframe like the model data frame with the same columns order,
field types, column names, and the "field_" attributes
}
