% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_s3_generics.R
\docType{methods}
\name{All generics}
\alias{All generics}
\alias{delete}
\alias{fetch}
\alias{fetch_vaults}
\title{deletes an object from EDP}
\usage{
delete(x, conn = get_connection())

fetch(x, conn = get_connection())

fetch_vaults(x, conn = get_connection())
}
\arguments{
\item{x}{the object to delete}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
the decorated object

the vault, or NULL if  not applicable
}
\description{
deletes an object from EDP

fetches an object using its ID

fetches the vault related to an object
}
