% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_connection}
\alias{get_connection}
\title{get the default connection if any}
\usage{
get_connection(auto = TRUE)
}
\arguments{
\item{auto}{whether to automatically use autoconnect() if the default connection is not yet set}
}
\description{
get the default connection if any
}
\seealso{
set_connection
}
