% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_url_to_parquet}
\alias{get_url_to_parquet}
\title{Get URL to parquet dataset}
\usage{
get_url_to_parquet(id, conn = get_connection())
}
\arguments{
\item{id}{(character) The ID of a QuartzBio EDP dataset, or a Dataset object.}

\item{conn}{(optional) Custom client environment.}
}
\value{
URL to the parquet file for the given EDP dataset. If there is an
failure, an error will be raised.
}
\description{
Get the associated URL to a parquet file for a given Quartzbio EDP dataset ID.
}
\concept{solvebio_api}
