% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_setup.R
\name{log_message}
\alias{log_message}
\title{Record a log message with the given log level}
\usage{
log_message(level, content)
}
\arguments{
\item{level}{log level Available levels: FATAL, ERROR, WARN, INFO, DEBUG}

\item{content}{message to be logged}
}
\description{
Add log messages with the given log level
}
\examples{
\dontrun{
log_message("INFO", "This is a info message")
log_message("WARN", "This is a warning")
log_message("ERROR", "This is an error")
log_message("FATAL", "this is a fatal message")
log_message("DEBUG", "this is a debug")
}
}
