% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{Multi_part_file_upload}
\alias{Multi_part_file_upload}
\title{Function to support multi-part upload for files larger than multipart_threshold.
Enhanced multipart upload with parallel parts and presigned URL refresh}
\usage{
Multi_part_file_upload(
  obj,
  local_path,
  local_md5,
  num_processes,
  max_retries,
  conn = get_connection(),
  ...
)
}
\arguments{
\item{obj}{Object created}

\item{local_path}{Path of file in local}

\item{local_md5}{md5sum checksum of file}

\item{num_processes}{Number of parallel workers for multipart upload (default: 1)}

\item{max_retries}{Maximum retries per part for multipart upload (default: 3)}

\item{conn}{Valid EDP connection}

\item{...}{
  Arguments passed on to \code{\link[=Objects]{Objects}}
  \describe{
    \item{\code{ancestor_id}}{an object ID of an ancestor, for filtering.}
    \item{\code{depth}}{the depth of the object in the Vault as an integer (0 means root)}
    \item{\code{filename}}{an Object filename, without the parent folder (e.g. "foo.csv")}
    \item{\code{glob}}{a glob (full path with wildcard characters) which searches
objects for matching paths (case-insensitive).}
    \item{\code{limit}}{The maximum number of elements to fetch, as an integer.
See also \code{page}.}
    \item{\code{min_distance}}{used in conjuction with the ancestor_id filter to only include
objects at a minimum distance from the ancestor.}
    \item{\code{object_type}}{the type of an object, one of  "file", "folder", "dataset" or "shortcut".}
    \item{\code{page}}{The number of the page to fetch, as an integer. starts from 1.
See also \code{limit}.}
    \item{\code{path}}{the path of an object, starting from the vault's
root directory. The path must include all intermediate folders and end with the object’s name
(e.g. "/d1/d2/foo.csv" or "/d1/d2/").}
    \item{\code{query}}{a string that matches any objects whose path contains that string.}
    \item{\code{regex}}{A regular expression, as a string, to filter the results with.}
    \item{\code{storage_class}}{The Storage class of the vault \verb{('Standard', 'Standard-IA', 'Essential', 'Temporary', 'Performance', 'Archive')} as a string.}
    \item{\code{tag}}{a single tag as a string.}
    \item{\code{vault_id}}{a Vault ID as a string (e.g. "19").}
    \item{\code{vault_name}}{a Vault name as a string (e.g. "Public").}
    \item{\code{vault_full_path}}{a Vault full path, as a string (e.g. "quartzbio:Public")}
  }}
}
\description{
Function to support multi-part upload for files larger than multipart_threshold.
Enhanced multipart upload with parallel parts and presigned URL refresh
}
