% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_object.R
\name{Object.all}
\alias{Object.all}
\title{Retrieves the metadata about all objects on EDP accessible to the current user.}
\usage{
Object.all(env = get_connection(), ...)
}
\arguments{
\item{env}{Custom client environment.}

\item{...}{Additional query parameters, passed to .request().}
}
\description{
Retrieves the metadata about all objects on EDP accessible to the current user.
}
\concept{solvebio_api}
