% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_object.R
\name{Object.disable_global_beacon}
\alias{Object.disable_global_beacon}
\title{Object.disable_global_beacon}
\usage{
Object.disable_global_beacon(id, env = get_connection())
}
\arguments{
\item{id}{The ID of a QuartzBio EDP dataset.}

\item{env}{(optional) Custom client environment.}
}
\description{
Disables Global Beacon for the specified dataset.
}
\examples{
\dontrun{
Object.disable_global_beacon("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{quartzbio_api}
