% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_object.R
\name{Object.fields}
\alias{Object.fields}
\title{Object.fields}
\usage{
Object.fields(id, env = get_connection(), ...)
}
\arguments{
\item{id}{The ID of a QuartzBio EDP file (vault object).}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. limit, offset).}
}
\description{
Retrieves the list of fields for a file (JSON, CSV, or TSV).
}
\examples{
\dontrun{
Object.fields("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{solvebio_api}
