% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_object.R
\name{Object.query}
\alias{Object.query}
\title{Object.query}
\usage{
Object.query(id, paginate = FALSE, env = get_connection(), ...)
}
\arguments{
\item{id}{The ID of a QuartzBio EDP file (vault object).}

\item{paginate}{When set to TRUE, retrieves all records (memory permitting).}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. filters, limit, offset).}
}
\description{
Queries a QuartzBio EDP file (vault object) and returns an R data frame containing all records.
Returns a single page of results otherwise (default).
}
\examples{
\dontrun{
Object.query("12345678790", paginate = TRUE)
}

}
\concept{quartzbio_api}
