% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object}
\alias{Object}
\title{fetches an object.}
\usage{
Object(
  id = NULL,
  full_path = NULL,
  path = NULL,
  vault_id = NULL,
  object_type = NULL,
  conn = get_connection()
)
}
\arguments{
\item{id}{an Object ID}

\item{full_path}{an Object full path, including the account, vault and path.}

\item{path}{the path of an object, starting from the vault's
root directory. The path must include all intermediate folders and end with the object’s name
(e.g. "/d1/d2/foo.csv" or "/d1/d2/").}

\item{vault_id}{a Vault ID as a string (e.g. "19").}

\item{object_type}{the type of an object, one of  "file", "folder", "dataset" or "shortcut".}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\description{
fetches an object.
}
