% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.retrieve}
\alias{Object.retrieve}
\title{Object.retrieve}
\usage{
Object.retrieve(id, env = quartzbio.edp:::.config)
}
\arguments{
\item{id}{The ID of the object.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific object on QuartzBio EDP.
}
\examples{
\dontrun{
Object.retrieve("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
