% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object_create}
\alias{Object_create}
\title{creates an Object.}
\usage{
Object_create(
  vault_id,
  filename,
  object_type,
  parent_object_id = NULL,
  description = NULL,
  metadata = NULL,
  tags = NULL,
  storage_class = NULL,
  capacity = NULL,
  mimetype = NULL,
  size = NULL,
  md5 = NULL,
  conn = get_connection()
)
}
\arguments{
\item{vault_id}{a Vault ID as a string (e.g. "19").}

\item{filename}{an Object filename, without the parent folder (e.g. "foo.csv")}

\item{object_type}{the type of an object, one of  "file", "folder", or "dataset".}

\item{parent_object_id}{the ID of the parent of the Object.}

\item{description}{the description as a string.}

\item{metadata}{metadata as a named list.}

\item{tags}{a list of tags as a character vector.}

\item{storage_class}{The Storage class of the vault \verb{('Standard', 'Standard-IA', 'Essential', 'Temporary', 'Performance', 'Archive')} as a string.}

\item{capacity}{The dataset capacity level (small, medium, or large).}

\item{mimetype}{the MIME type of the Object.}

\item{size}{The size of the object.}

\item{md5}{a MD5 fingerprint, as a string.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
the object as as list with class Object
}
\description{
creates an Object.
}
