% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object_update}
\alias{Object_update}
\title{updates an Object.}
\usage{
Object_update(
  id,
  filename = NULL,
  object_type = NULL,
  parent_object_id = NULL,
  description = NULL,
  metadata = NULL,
  tags = NULL,
  storage_class = NULL,
  conn = get_connection()
)
}
\arguments{
\item{id}{an Object ID}

\item{filename}{an Object filename, without the parent folder (e.g. "foo.csv")}

\item{object_type}{the type of an object, one of  "file", "folder", "dataset" or "shortcut".}

\item{parent_object_id}{the ID of the parent of the Object.}

\item{description}{the description as a string.}

\item{metadata}{metadata as a named list.}

\item{tags}{a list of tags as a character vector.}

\item{storage_class}{The Storage class of the vault \verb{('Standard', 'Standard-IA', 'Essential', 'Temporary', 'Performance', 'Archive')} as a string.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
the object as as list with class Object
}
\description{
updates an Object.
}
