% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_shared_params.R
\name{old_params}
\alias{old_params}
\title{shared old roxygen params}
\arguments{
\item{client_id}{The client ID for the application.}

\item{conn}{a EDP connection object (as a named list or environment)}

\item{data}{TODO}

\item{data_type}{The data type to cast the expression result.}

\item{env}{Custom client environment.}

\item{expression}{The EDP expression string.}

\item{fields}{The fields to add.}

\item{include_errors}{whether to include errors in the output.}

\item{is_list}{whether the result is expected to be a list.}

\item{raw}{whether to return the raw response.}

\item{records}{The data to annotate as a data frame.}

\item{...}{Additional query parameters, passed to .request().

\url{https://docs.solvebio.com/}}
}
\description{
shared old roxygen params
}
