% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_shiny.R
\name{quartzbio_shiny_auth}
\alias{quartzbio_shiny_auth}
\title{New function for EDP auth of shiny app. Returns the
updated session with EDP connection to be used in shiny app}
\usage{
quartzbio_shiny_auth(input, session, client_id, client_secret = NULL, base_url)
}
\arguments{
\item{input}{input}

\item{session}{session}

\item{client_id}{client_id}

\item{client_secret}{secret}

\item{base_url}{BASE URL}
}
\description{
New function for EDP auth of shiny app. Returns the
updated session with EDP connection to be used in shiny app
}
\concept{quartzbio_api}
