% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{read_connection_profile}
\alias{read_connection_profile}
\title{read a connection profile}
\usage{
read_connection_profile(
  profile = get_env("EDP_PROFILE", "default"),
  path = get_env("EDP_CONFIG", "~/.qb/edp.json")
)
}
\arguments{
\item{profile}{the name of the profile, as a string}

\item{path}{the path to the connection profiles file, as a string}
}
\value{
the connection for the given profile as a named list, or die if there is no such profile
}
\description{
read a connection profile
}
\seealso{
Other connection: 
\code{\link{save_connection_profile}()}
}
\concept{connection}
