% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{refresh_presigned_urls}
\alias{refresh_presigned_urls}
\title{Refresh presigned URLs for multipart upload}
\usage{
refresh_presigned_urls(
  upload_id,
  key,
  size,
  part_numbers,
  conn = get_connection(),
  ...
)
}
\arguments{
\item{upload_id}{description}

\item{key}{description}

\item{size}{description}

\item{part_numbers}{description}

\item{conn}{Valid EDP connection}

\item{...}{
  Arguments passed on to \code{\link[=Objects]{Objects}}
  \describe{
    \item{\code{ancestor_id}}{an object ID of an ancestor, for filtering.}
    \item{\code{depth}}{the depth of the object in the Vault as an integer (0 means root)}
    \item{\code{filename}}{an Object filename, without the parent folder (e.g. "foo.csv")}
    \item{\code{glob}}{a glob (full path with wildcard characters) which searches
objects for matching paths (case-insensitive).}
    \item{\code{limit}}{The maximum number of elements to fetch, as an integer.
See also \code{page}.}
    \item{\code{min_distance}}{used in conjuction with the ancestor_id filter to only include
objects at a minimum distance from the ancestor.}
    \item{\code{object_type}}{the type of an object, one of  "file", "folder", "dataset" or "shortcut".}
    \item{\code{page}}{The number of the page to fetch, as an integer. starts from 1.
See also \code{limit}.}
    \item{\code{path}}{the path of an object, starting from the vault's
root directory. The path must include all intermediate folders and end with the object’s name
(e.g. "/d1/d2/foo.csv" or "/d1/d2/").}
    \item{\code{query}}{a string that matches any objects whose path contains that string.}
    \item{\code{regex}}{A regular expression, as a string, to filter the results with.}
    \item{\code{storage_class}}{The Storage class of the vault \verb{('Standard', 'Standard-IA', 'Essential', 'Temporary', 'Performance', 'Archive')} as a string.}
    \item{\code{tag}}{a single tag as a string.}
    \item{\code{vault_id}}{a Vault ID as a string (e.g. "19").}
    \item{\code{vault_name}}{a Vault name as a string (e.g. "Public").}
    \item{\code{vault_full_path}}{a Vault full path, as a string (e.g. "quartzbio:Public")}
  }}
}
\description{
Refresh presigned URLs for multipart upload
}
