% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_saved_query.R
\name{SavedQuery.all}
\alias{SavedQuery.all}
\title{SavedQuery.all}
\usage{
SavedQuery.all(env = get_connection(), ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. page).}
}
\description{
Retrieves the all saved queries on QuartzBio EDP.
}
\examples{
\dontrun{
SavedQuery.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{quartzbio_api}
