% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_saved_query.R
\name{SavedQuery.create}
\alias{SavedQuery.create}
\title{SavedQuery.create}
\usage{
SavedQuery.create(env = get_connection(), ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Saved query attributes.}
}
\description{
Create a QuartzBio EDP saved query.
}
\examples{
\dontrun{
SavedQuery.create(name="My Dataset Template")
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{solvebio_api}
