% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saved_query.R
\name{SavedQuery.delete}
\alias{SavedQuery.delete}
\title{SavedQuery.delete}
\usage{
SavedQuery.delete(id, env = quartzbio.edp:::.config)
}
\arguments{
\item{id}{String The ID of the QuartzBio EDP saved query.}

\item{env}{(optional) Custom client environment.}
}
\description{
Deletes a specific saved query on QuartzBio EDP.
}
\examples{
\dontrun{
SavedQuery.delete(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
