% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_saved_query.R
\name{SavedQuery.update}
\alias{SavedQuery.update}
\title{SavedQuery.update}
\usage{
SavedQuery.update(id, env = get_connection(), ...)
}
\arguments{
\item{id}{The ID of the saved query to update.}

\item{env}{(optional) Custom client environment.}

\item{...}{Saved query attributes to change.}
}
\description{
Updates the attributes of an existing saved query.
}
\examples{
\dontrun{
SavedQuery.update(
  id = "1234",
  name = "New query Name",
)
}

}
\concept{quartzbio_api}
