% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{set_connection}
\alias{set_connection}
\title{set the default connection}
\usage{
set_connection(conn, check = TRUE)
}
\arguments{
\item{conn}{a EDP connection object (as a named list or environment)}

\item{check}{whether to check the connection, mostly for debugging purposes}
}
\description{
N.B: use conn=NULL to unset the default connection
may die on bad connection
}
