% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut.R
\name{Shortcut}
\alias{Shortcut}
\title{fetches a shortcut by id or (vault_id, path)}
\usage{
Shortcut(id = NULL, path = NULL, vault_id = NULL, conn = get_connection())
}
\arguments{
\item{id}{the shortcut ID.}

\item{path}{the path of an object, starting from the vault's
root directory. The path must include all intermediate folders and end with the object’s name
(e.g. "/d1/d2/foo.csv" or "/d1/d2/").}

\item{vault_id}{a Vault ID as a string (e.g. "19").}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\description{
fetches a shortcut by id or (vault_id, path)
}
