% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut.R
\name{Shortcut_create}
\alias{Shortcut_create}
\title{creates  a shortcut.}
\usage{
Shortcut_create(
  vault_id,
  vault_path,
  target,
  tags = list(),
  conn = get_connection()
)
}
\arguments{
\item{vault_id}{a Vault ID as a string (e.g. "19").}

\item{vault_path}{a Vault path, as a string (e.g. "/d1/d2/foo.csv")}

\item{target}{named list. Only relevant when \code{object_type} is \code{shortcut}.
Must contain \code{object_type}. When \code{object_type} is \code{url} then the second field
must be \code{url}. For other \code{object_type}s \code{id} should be provided.
Shortcuts can be created to: \itemize{
\item vaults (object_type='vault')
\item files (object_type='file')
\item datasets (object_type='dataset')
\item folders (object_type='folder')
\item URLs (object_type='url')
}}

\item{tags}{a list of tags as a character vector.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
the shortcut as an Object
}
\description{
creates  a shortcut.
}
\details{
Shortcuts cannot be created to: \itemize{
\item the containing vault of the shortcut file.
\item the parent folder or any ancestor of the shortcut file.
\item a deleted object
\item a non-existent object
\item vaults that are not of type 'general'}
}
