% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut.R
\name{Shortcut_get_target}
\alias{Shortcut_get_target}
\title{Get Shortcut target}
\usage{
Shortcut_get_target(obj, allow_null_target = TRUE)
}
\arguments{
\item{obj}{\code{Object}}

\item{allow_null_target}{Bool (Optional). If False will raise an error instead of returning NULL}
}
\value{
Object of type \code{Vault} or \code{Object} or \code{character} or \code{NULL} depending on the target
}
\description{
Resolves the target of a shortcut into its object type.
}
