% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{Task}
\alias{Task}
\title{fetches a task.}
\usage{
Task(task_id, conn = get_connection())
}
\arguments{
\item{task_id}{an (ECS) Task ID as a string.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
the tasks as a ECSTask object, or NULL if none found.
}
\description{
fetches a task.
}
