% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{Task_wait_for_completion}
\alias{Task_wait_for_completion}
\title{Waits for a task to be completed (or failed).}
\usage{
Task_wait_for_completion(
  task_id,
  interval = 3,
  retries = 30,
  recursive = TRUE,
  conn = get_connection()
)
}
\arguments{
\item{task_id}{an (ECS) Task ID as a string.}

\item{interval}{time in seconds to wait before retrying}

\item{retries}{number of attempts to perform to check the task completion}

\item{recursive}{whether to also wait for the subtasks}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
TRUE if the task is finished (completed or failed), or FALSE if the number of retries
is exceeded (~ timeout).
}
\description{
N.B: using retries == 0 will immediately timeout and return FALSE. This is convenient for
unit testing for example
}
