% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{test_connection}
\alias{test_connection}
\title{test a connection}
\usage{
test_connection(conn)
}
\arguments{
\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
TRUE iff the connection was successful
}
\description{
may die on bad connection
}
