% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{User}
\alias{User}
\title{fetches  user information.}
\usage{
User(conn = get_connection())
}
\arguments{
\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
the connected user information as a User object
}
\description{
fetches  user information.
}
