% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{User.retrieve}
\alias{User.retrieve}
\title{Retrieves information about the current user}
\usage{
User.retrieve(env = get_connection())
}
\arguments{
\item{env}{Custom client environment.}
}
\description{
Retrieves information about the current user
}
