% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_vault.R
\name{Vault.delete}
\alias{Vault.delete}
\title{Vault.delete}
\usage{
Vault.delete(id, env = get_connection())
}
\arguments{
\item{id}{String The ID of a QuartzBio EDP vault.}

\item{env}{(optional) Custom client environment.}
}
\description{
Delete a specific vault from QuartzBio EDP. This operation cannot be undone.
}
\examples{
\dontrun{
Vault.delete("1")
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{quartzbio_api}
