% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault}
\alias{Vault}
\title{fetches a vault}
\usage{
Vault(id = NULL, full_path = NULL, name = NULL, conn = get_connection())
}
\arguments{
\item{id}{the Vault ID or object to fetch}

\item{full_path}{the full path of the vault to fetch}

\item{name}{the name of the the vault to fetch}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
the vault as a list with class \code{Vault}, or NULL if no matching vault is found
}
\description{
N.B:
\itemize{
\item if called without \code{id}, \code{full_path} or \code{name}, it fetches the \emph{personal vault}
of the connected user.
\item dies if multiple vaults are matched
}
}
\examples{
\dontrun{
# with no argument, fetch the connected user personal vault
v <- Vault()

# by id
v2 <- Vault(v$id)

# by full_path
v2 <- Vault(full_path = v$full_path)

# by name
v2 <- Vault(name = "Public")
}
}
