% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_vault.R
\name{Vault.retrieve}
\alias{Vault.retrieve}
\title{Vault.retrieve}
\usage{
Vault.retrieve(id, env = get_connection())
}
\arguments{
\item{id}{String The ID of a QuartzBio EDP vault}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific QuartzBio EDP vault.
}
\examples{
\dontrun{
Vault.retrieve("1234567890")
}

}
\concept{quartzbio_api}
