% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_vault.R
\name{Vault.update}
\alias{Vault.update}
\title{Vault.update}
\usage{
Vault.update(id, env = get_connection(), ...)
}
\arguments{
\item{id}{The ID of the vault to update.}

\item{env}{(optional) Custom client environment.}

\item{...}{Vault attributes to change.}
}
\description{
Updates the attributes of an existing vault.
}
\examples{
\dontrun{
Vault.update(
  id = "1234",
  name = "New Vault Name",
)
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{solvebio_api}
