% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault_create}
\alias{Vault_create}
\title{creates a new EDP vault or updates an existing one.}
\usage{
Vault_create(
  name,
  description = NULL,
  metadata = NULL,
  tags = NULL,
  storage_class = NULL,
  conn = get_connection()
)
}
\arguments{
\item{name}{the vault name to create, as a string.}

\item{description}{the description as a string.}

\item{metadata}{metadata as a named list.}

\item{tags}{a list of tags as a character vector.}

\item{storage_class}{The Storage class of the vault \verb{('Standard', 'Standard-IA', 'Essential', 'Temporary', 'Performance', 'Archive')} as a string.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\description{
cf https://docs.solvebio.com/reference/vaults/vaults/#create
}
\details{
if id is NULL, it will create a new vault, otherwise it will update
the vault with the corresponding id
}
\examples{
\dontrun{
# simplest form
v <- Vault_create("my.new.vault")

# using all params
v <- Vault_create(
  name = "my.new.vault",
  description = "This is my own vault",
  metadata = list(a = 1, b = "toto", sublist = list(x = "str")),
  tags = c("TEST", "DATA"),
  storage_class = "Performance",
  conn = conn
)
}
}
