% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault_update}
\alias{Vault_update}
\title{updates a vault}
\usage{
Vault_update(
  id,
  name = NULL,
  description = NULL,
  metadata = NULL,
  tags = NULL,
  storage_class = NULL,
  conn = get_connection()
)
}
\arguments{
\item{id}{the Vault ID or object to fetch}

\item{name}{the vault name to create, as a string.}

\item{description}{the description as a string.}

\item{metadata}{metadata as a named list.}

\item{tags}{a list of tags as a character vector.}

\item{storage_class}{The Storage class of the vault \verb{('Standard', 'Standard-IA', 'Essential', 'Temporary', 'Performance', 'Archive')} as a string.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\value{
a Vault object
}
\description{
N.B:
\itemize{
\item the updated vault properties are overwritten, not merged!
\item if \code{id} is NULL, this will create a new vault
}
}
\examples{
\dontrun{
  v2 <- Vault_update(v$id, name = name, description = 'desc',    
   metadata = list(meta1 = 'toto'), storage_class = 'Performance', tags = 'A')
# using methods
v3 <- update(v2, name = name, storage_class = 'Temporary')
v4 <- update(vault_id, tags = LETTERS[1:5])
}
}
