% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vaults}
\alias{Vaults}
\title{fetches a list of vaults}
\usage{
Vaults(
  vault_type = NULL,
  tags = NULL,
  user_id = NULL,
  storage_class = NULL,
  account_id = NULL,
  page = NULL,
  limit = NULL,
  conn = get_connection()
)
}
\arguments{
\item{vault_type}{the type of vault ('user', 'general') as a string}

\item{tags}{a list of tags as a character vector.}

\item{user_id}{a user id (or User object) as a string}

\item{storage_class}{The Storage class of the vault \verb{('Standard', 'Standard-IA', 'Essential', 'Temporary', 'Performance', 'Archive')} as a string.}

\item{account_id}{an Account ID as a string.}

\item{page}{The number of the page to fetch, as an integer. starts from 1.
See also \code{limit}.}

\item{limit}{The maximum number of elements to fetch, as an integer.
See also \code{page}.}

\item{conn}{a EDP connection object (as a named list or environment)}
}
\description{
fetches a list of vaults
}
