% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_application.R
\name{Application.retrieve}
\alias{Application.retrieve}
\title{Retrieve the metadata about a specific application QuartzBio EDP.}
\usage{
Application.retrieve(client_id, env = get_connection())
}
\arguments{
\item{client_id}{the client ID for the application.}

\item{env}{Custom client environment.}
}
\description{
Retrieve the metadata about a specific application QuartzBio EDP.
}
\examples{
\dontrun{
Application.retrieve("abcd1234")
}
}
\concept{quartzbio_api}
