% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_beacon.R
\name{Beacon.retrieve}
\alias{Beacon.retrieve}
\title{Beacon.retrieve}
\usage{
Beacon.retrieve(id, env = get_connection())
}
\arguments{
\item{id}{The ID of the beacon.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific beacon on QuartzBio EDP.
}
\examples{
\dontrun{
Beacon.retrieve("1234")
}

}
\concept{quartzbio_api}
