% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset_load}
\alias{Dataset_load}
\title{Dataset_load}
\usage{
Dataset_load(id = NULL, full_path = NULL, get_schema = FALSE, ...)
}
\arguments{
\item{id}{(character) The ID of a QuartzBio EDP dataset.}

\item{full_path}{(character) a valid dataset full path, including the account, vault and path to EDP Dataset.}

\item{get_schema}{(boolean) Retrieves the schema of the Quartzbio EDP dataset loaded. Default value: FALSE}

\item{...}{
  Arguments passed on to \code{\link[arrow:read_parquet]{arrow::read_parquet}}
  \describe{
    \item{\code{col_select}}{A character vector of column names to keep, as in the
"select" argument to \code{data.table::fread()}, or a
\link[tidyselect:eval_select]{tidy selection specification}
of columns, as used in \code{dplyr::select()}.}
    \item{\code{as_data_frame}}{Should the function return a \code{tibble} (default) or
an Arrow \link[arrow]{Table}?}
  }}
}
\value{
A \code{tibble} which is the default, or an Arrow Table otherwise. If the \code{get_schema} parameter is set to \code{TRUE},
the function returns a list containing both the \code{tibble} and its schema.
}
\description{
Loads large Quartzbio EDP dataset and returns an R data frame containing all records.
}
\concept{quartzbio_api}
