% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset_schema}
\alias{Dataset_schema}
\title{Get dataset schema}
\usage{
Dataset_schema(id)
}
\arguments{
\item{id}{The ID of a QuartzBio EDP dataset.}
}
\value{
A Schema object containing Fields, which maps to the data types.
}
\description{
Retrieves the schema of the Quartzbio EDP dataset.
}
\concept{solvebio_api}
