% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset_schema}
\alias{Dataset_schema}
\title{Get dataset schema}
\usage{
Dataset_schema(id = NULL, full_path = NULL)
}
\arguments{
\item{id}{The ID of a QuartzBio EDP dataset.}

\item{full_path}{a valid dataset full path, including the account, vault and path to EDP Dataset.}
}
\value{
A Schema object containing Fields, which maps to the data types.
}
\description{
Retrieves the schema of the Quartzbio EDP dataset.
}
\concept{solvebio_api}
