% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset_export.R
\name{DatasetExport.retrieve}
\alias{DatasetExport.retrieve}
\title{DatasetExport.retrieve}
\usage{
DatasetExport.retrieve(id, env = get_connection())
}
\arguments{
\item{id}{String The ID of a QuartzBio EDP dataset export.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific dataset export on QuartzBio EDP.
}
\examples{
\dontrun{
DatasetExport.retrieve(12345)
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{solvebio_api}
