% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dataset_migration.R
\name{DatasetMigration.retrieve}
\alias{DatasetMigration.retrieve}
\title{DatasetMigration.retrieve}
\usage{
DatasetMigration.retrieve(id, env = get_connection())
}
\arguments{
\item{id}{String The ID of a QuartzBio EDP dataset migration.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific dataset migration on QuartzBio EDP.
}
\examples{
\dontrun{
DatasetMigration.retrieve(12345)
}

}
\references{
\url{https://docs.solvebio.com/}
}
\concept{solvebio_api}
