% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{login}
\usage{
login(
  api_key = Sys.getenv("SOLVEBIO_API_KEY"),
  api_token = Sys.getenv("SOLVEBIO_ACCESS_TOKEN"),
  api_host = Sys.getenv("SOLVEBIO_API_HOST"),
  env = get_connection()
)
}
\arguments{
\item{api_key}{Your QuartzBio EDP API key}

\item{api_host}{QuartzBio EDP API host (default: https://api.solvebio.com)}

\item{env}{(optional) The R environment used to store API credentials.}
}
\description{
Store and verify your QuartzBio EDP credentials.
}
\examples{
\dontrun{
login()
}

}
\references{
\url{https://docs.solvebio.com/}
}
