% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_vault.R
\name{Vault.create}
\alias{Vault.create}
\title{Vault.create}
\usage{
Vault.create(name, env = get_connection(), ...)
}
\arguments{
\item{name}{The unique name of the vault.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional vault attributes.}
}
\description{
Create a new QuartzBio EDP vault.
}
\examples{
\dontrun{
Vault.create(name = "my-domain:MyVault")
}

}
\concept{quartzbio_api}
